fix: use correct license (#4)
This code was taken from the Node.js core repository and one of the
few conditions of that repository’s MIT license is that the MIT
license text needs to be preserved. We cannot pretend to re-license
it as Apache 2, even if there may not be a lot of practical difference
for consumers.